+++ src/bluesearch/entrypoint/database/topic_extract.py
@@ -21,7 +21,7 @@ import argparse
 import datetime
 import logging
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any
 
 logger = logging.getLogger(__name__)
 
@@ -117,7 +117,7 @@ def run(
     source: str,
     input_path: Path,
     output_file: Path,
-    match_filename: Optional[str],
+    match_filename: str | None,
     recursive: bool,
     overwrite: bool,
     dry_run: bool,
@@ -127,9 +127,13 @@ def run(
     Parameter description and potential defaults are documented inside of the
     `get_parser` function.
     """
+    from defusedxml import ElementTree
+
     import bluesearch
     from bluesearch.database.topic import (
+        extract_article_topics_for_pubmed_article,
         extract_info_from_zipfile,
+        extract_journal_topics_for_pubmed_article,
         get_topics_for_pmc_article,
     )
     from bluesearch.utils import JSONL, find_files
@@ -195,6 +199,35 @@ def run(
             )
 
         pass
+
+    elif source == "pubmed":
+        for path in inputs:
+            logger.info(f"Processing {path}")
+            articles = ElementTree.parse(input_path)
+            for i, article in enumerate(articles.iter("PubmedArticle")):
+                article_topics = extract_article_topics_for_pubmed_article(article)
+                journal_topics = extract_journal_topics_for_pubmed_article(article)
+                all_results.append(
+                    {
+                        "source": "pubmed",
+                        "path": str(path.resolve()),
+                        "topics": {
+                            "journal": {
+                                "MeSH": journal_topics,
+                            },
+                            "article": {
+                                "MeSH": article_topics,
+                            },
+                        },
+                        "metadata": {
+                            "created-date": datetime.datetime.now().strftime(
+                                "%Y-%m-%d %H:%M:%S"
+                            ),
+                            "bbs-version": bluesearch.version.__version__,
+                            "element_in_file": i,
+                        },
+                    }
+                )
     else:
         logger.error(f"The source type {source!r} is not implemented yet")
         return 1